requirements.txt
astroid==2.3.1
certifi==2019.9.11
chardet==3.0.4
Django==2.2.6
django-countries==5.5
django-oauth-toolkit==1.2.0
djangorestframework==3.10.3
idna==2.8
isort==4.3.21
lazy-object-proxy==1.4.2
mccabe==0.6.1
oauthlib==3.1.0
Pillow==6.2.0
pycodestyle==2.5.0
pydocstyle==4.0.1
pyflakes==2.1.1
pylama==7.7.1
pylint==2.4.2
pylint-django==2.0.11
pylint-plugin-utils==0.6
pytz==2019.2
requests==2.22.0
six==1.12.0
snowballstemmer==2.0.0
sqlparse==0.3.0
typed-ast==1.4.0
urllib3==1.25.6
wrapt==1.11.2
